requirements.txt
pyserial
pygame
matplotlib